content/html/content/test/test_video_wakelock.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=868943
-->
<head>
  <title>Test for Bug 868943</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=868943">Mozilla Bug 868943</a>
<p id="display"></p>
<div id="content">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 868943 **/

SpecialPowers.addPermission("power", true, document);

function testVideoPlayPause() {
  var lockState = true;
  var count = 0;

  var content = document.getElementById('content');

  var video = document.createElement('video');
  video.src = "wakelock.ogv";
  content.appendChild(video);

  navigator.mozPower.addWakeLockListener(function testVideoPlayPauseListener(topic, state) {
    is(topic, "screen", "Video element locked the target == screen");
    var locked = state == "locked-foreground" ||
                 state == "locked-background";

    is(locked, lockState, "Video element locked the screen - paused");
    count++;

    if (count == 1) {
      // The next step is to unlock the resource.
      lockState = false;
      video.pause();
    } else if (count == 2) {
      content.removeChild(video);
      navigator.mozPower.removeWakeLockListener(testVideoPlayPauseListener);
      runTests();
    }
  });

  video.play();
}

function testVideoPlay() {
  var lockState = true;
  var count = 0;

  var content = document.getElementById('content');

  var video = document.createElement('video');
  video.src = "wakelock.ogv";
  content.appendChild(video);

  navigator.mozPower.addWakeLockListener(function testVideoPlayListener(topic, state) {
    is(topic, "screen", "Video element locked the target == screen");
    var locked = state == "locked-foreground" ||
                 state == "locked-background";

    is(locked, lockState, "Video element locked the screen - no paused");
    count++;

    if (count == 1) {
      // The next step is to unlock the resource.
      lockState = false;
    } else if (count == 2) {
      content.removeChild(video);
      navigator.mozPower.removeWakeLockListener(testVideoPlayListener);
      runTests();
    }
  });

  video.play();
}

var tests = [ testVideoPlayPause, testVideoPlay ];
function runTests() {
  if (!tests.length) {
    SimpleTest.finish();
    return;
  }

  var test =  tests.pop();
  test();
};

SimpleTest.waitForExplicitFinish();
runTests();

</script>
</pre>
</body>
</html>
